Add support for big files (lager than 2GB)

+++ apt-pkg/acquire-item.h
@@ -45,8 +45,8 @@ class pkgAcquire::Item
     * Do not use anonyomus enums, as this breaks SWIG in some cases */
    enum StatusFlags {StatIdle, StatFetching, StatDone, StatError} Status;
    string ErrorText;
-   unsigned long FileSize;
-   unsigned long PartialSize;   
+   unsigned long long FileSize;
+   unsigned long long PartialSize;   
    const char *Mode;
    unsigned long ID;
    bool Complete;
@@ -60,9 +60,9 @@ class pkgAcquire::Item
 
    // Action members invoked by the worker
    virtual void Failed(string Message,pkgAcquire::MethodConfig *Cnf);
-   virtual void Done(string Message,off_t Size,string Md5Hash,
+   virtual void Done(string Message,unsigned long long Size,string Md5Hash,
 		     pkgAcquire::MethodConfig *Cnf);
-   virtual void Start(string Message,unsigned long Size);
+   virtual void Start(string Message,unsigned long long Size);
    virtual string Custom600Headers() {return string();}
    virtual string DescURI() = 0;
    virtual void Finished() {}
@@ -97,7 +97,7 @@ class pkgAcqIndex : public pkgAcquire::Item
    public:
    
    // Specialized action members
-   virtual void Done(string Message,off_t Size,string Md5Hash,
+   virtual void Done(string Message,unsigned long long Size,string Md5Hash,
 		     pkgAcquire::MethodConfig *Cnf);
    virtual string Custom600Headers();
    virtual string DescURI() {return RealURI;} // CNC:2003-02-14
@@ -128,7 +128,7 @@ class pkgAcqIndexRel : public pkgAcquire::Item
    
    // Specialized action members
    virtual void Failed(string Message,pkgAcquire::MethodConfig *Cnf);
-   virtual void Done(string Message,off_t Size,string Md5Hash,
+   virtual void Done(string Message,unsigned long long Size,string Md5Hash,
 		     pkgAcquire::MethodConfig *Cnf);   
    virtual string Custom600Headers();
    virtual string DescURI() {return RealURI;}
@@ -161,7 +161,7 @@ class pkgAcqArchive : public pkgAcquire::Item
    
    // Specialized action members
    virtual void Failed(string Message,pkgAcquire::MethodConfig *Cnf);
-   virtual void Done(string Message,off_t Size,string Md5Hash,
+   virtual void Done(string Message,unsigned long long Size,string Md5Hash,
 		     pkgAcquire::MethodConfig *Cnf);
    virtual string MD5Sum() {return MD5;}
    virtual string DescURI() {return Desc.URI;}
@@ -186,12 +186,12 @@ class pkgAcqFile : public pkgAcquire::Item
    
    // Specialized action members
    virtual void Failed(string Message,pkgAcquire::MethodConfig *Cnf);
-   virtual void Done(string Message,off_t Size,string Md5Hash,
+   virtual void Done(string Message,unsigned long long Size,string Md5Hash,
 		     pkgAcquire::MethodConfig *Cnf);
    virtual string MD5Sum() {return Md5Hash;}
    virtual string DescURI() {return Desc.URI;}
    
-   pkgAcqFile(pkgAcquire *Owner,string URI,string MD5,unsigned long Size,
+   pkgAcqFile(pkgAcquire *Owner,string URI,string MD5,unsigned long long Size,
 		  string Desc,string ShortDesc);
 };
